requirements.txt
langchain
openai
jupyter
streamlit
python-dotenv
streamlit-chat
